Reorganize validation
